Move Vomit related classes to mobs.py
